Fetch a random article
ArticleContainer can now fetch a random article upon first mounting.

A random integer between 1 and 5 is generated and passed to the api get request. Each number sent to the backend requests a JSON article object associated with that number.

For the time being, we can fetch the same article twice or more but this will be fixed later on where we will fetch only one article at a time and not have it fetched again