Fetch next article and readArticles feature
Code is initially a bit dirty but the gist of it is it can now fetch not only the current article data but also the upcoming article. This will allow the user a seamless navigation experience (the next article is pre-loaded as the user reads the current article) rather than fetching a new current article each time the user navigates to a different article page.