Add Rankings feature
Currently, as a user navigates through 4 random articles, a Proceed To Rankings button appears which when clicked, navigates the user to a Rankings component. The Rankings component takes in the parent readArticles array as props and displays each item of that array with their own star rating system.

The user can rank each article that was read by rating each listed article title. The Rankings component makes sure the user ranked all 5 articles before allowing them to post the crafted data (an array of objects, where each object has a rating and a article title property) to a backend.

At present no messages giving feedback to the user is in place but will be implemented next.