Generate random number once +Show user feedback
Using a generator in helpers.js, we can now generate a number only once, meaning we only pass a non-repeating number as query parameter to the http get request which means we will never fetch the same article data more than once like before

Messages are also displayed if the app detects an offline connection or a server status code while fetching an article. Appropriate error messages are displayed if a user cannot post the rankings data to the backend. A success message is displayed if the user can, however.